+++ src/Type/WPInterface/NodeWithSeo.php
@@ -16,6 +16,8 @@ use WPGraphQL\RankMath\Model\ContentNodeSeo;
 use WPGraphQL\RankMath\Model\ContentTypeSeo;
 use WPGraphQL\RankMath\Model\TermNodeSeo;
 use WPGraphQL\RankMath\Model\UserSeo;
+use WPGraphQL\RankMath\Type\WPInterface\ContentNodeSeo as WPInterfaceContentNodeSeo;
+use WPGraphQL\RankMath\Utils\Utils;
 use WPGraphQL\RankMath\Vendor\AxeWP\GraphQL\Abstracts\InterfaceType;
 use WPGraphQL\RankMath\Vendor\AxeWP\GraphQL\Interfaces\TypeWithInterfaces;
 
@@ -50,6 +52,11 @@ class NodeWithSeo extends InterfaceType implements TypeWithInterfaces {
 		// @todo only apply to ContentTypes that have SEO data.
 
 		register_graphql_interfaces_to_types( self::type_name(), $types_with_seo );
+
+		// Narrow down ContentNode types.
+		Utils::overload_graphql_field_type( 'ContentNode', 'seo', WPInterfaceContentNodeSeo::get_type_name() );
+		// This is necessary because the filter doesn't work for inheritance.
+		Utils::overload_graphql_field_type( 'HierarchicalContentNode', 'seo', WPInterfaceContentNodeSeo::get_type_name() );
 	}
 
 	/**